requirements.txt
flask
python-keystoneclient
mariadb